requirements.txt
backports.zoneinfo==0.2.1
click==8.1.0
colorama==0.4.4
Flask==2.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg==3.0.11
psycopg-binary==3.0.11
tzdata==2022.1
Werkzeug==2.1.0
zipp==3.7.0
